# refactor/perf(factory): optimize message/field lookup, reduce memory