Previous lookup structure take a lot of space: `4.345.200 KB (4.3 MB)`. This PR simplify the lookup structure to only include necessary data. Actually, we don't need to have a full Field structure, we only need the FieldBase, so we create an array that consist only the FieldBase and create the Field accordingly on the `CreateField` method. This saves a lot of space, it is now only takes `390.512 KB`!

Since we don't have a full Field with its respective invalid value like previous implementation, Field created by `CreateField` method is now returns with `proto.Value` with `proto.TypeInvalid`. We believe it's okay to make this trade-off, since the Field's Value is expected to be filled by users on manual creation, even if they don't, by default, the invalid values will be omitted by Encoder anyway. Messages created by Decoder will always have Fields with valid value.

This is how we measure memory usage for the previous structure and the current structure respectively:
```go
totalSize := unsafe.Sizeof(mesgs) // var mesg = [...]message{Num uint16, Fields [256]proto.Field}
fmt.Printf("mesgs: %d\n", totalSize)
for _, mesg := range mesgs {
	for _, field := range mesg.Fields {
		if field.FieldBase != nil {
			totalSize += unsafe.Sizeof(*field.FieldBase)
		}
	}
}
fmt.Printf("total: %d\n", totalSize)

// Output:
// mesgs: 4201680
// total: 4345200
```

```go
totalSize := unsafe.Sizeof(mesgs) // var mesgs = [...]*[256]*proto.FieldBase{}
fmt.Printf("mesgs: %d\n", totalSize)
for _, fieldBases := range mesgs {
	if fieldBases == nil {
		continue
	}
	totalSize += unsafe.Sizeof(*fieldBases)
	for _, fieldBase := range fieldBases {
		if fieldBase != nil {
			totalSize += unsafe.Sizeof(*fieldBase)
		}
	}
}
fmt.Printf("total: %d\n", totalSize)

// Output:
// mesgs: 3280
// total: 390512
```

We don't see any performance degradation by implementing the new lookup structure, at least in our benchmark:
```js
goos: darwin
goarch: amd64
pkg: benchfit
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkDecode/muktihari/fit_raw-4  12    93768987 ns/op   77092821 B/op    100044 allocs/op
BenchmarkDecode/muktihari/fit-4      13    85676532 ns/op   52205447 B/op    100190 allocs/op
BenchmarkDecode/tormoder/fit-4       10   108446738 ns/op   84108974 B/op    700051 allocs/op
BenchmarkEncode/muktihari/fit_raw-4  19    58959556 ns/op     131418 B/op        11 allocs/op
BenchmarkEncode/muktihari/fit-4       9   118428917 ns/op   44138958 B/op    100014 allocs/op
BenchmarkEncode/tormoder/fit-4        1  1313825616 ns/op  102392720 B/op  12100314 allocs/op
PASS
ok      benchfit        12.572s
```